AI File Reader/backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic>=2.11.2,<3.0.0
requests==2.31.0
psycopg2-binary>=2.9.9,<2.10.0
clerk-backend-api==3.1.11
fastapi-clerk-auth==0.0.7
python-jose[cryptography]==3.3.0
openai>=1.13.3,<2.0.0
tiktoken==0.7.0
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-qdrant==0.1.4
qdrant-client>=1.10.1,<2.0.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.2
xlrd==2.0.1